Added files as per review with auto pep8 all files
