+++ gbpservice/neutron/nfp/config_agent/rpc_cb.py
@@ -1,12 +1,22 @@
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+
 import importlib
-
-from gbpservice.neutron.nsf.core import periodic_task as core_pt
-from gbpservice.neutron.nsf.config_agent import RestClientOverUnix as rc
-from gbpservice.neutron.nsf.config_agent import loadbalancer as lb
-from gbpservice.neutron.nsf.config_agent import firewall as fw
-from gbpservice.neutron.nsf.config_agent import vpn as vpn
-from gbpservice.neutron.nsf.config_agent import generic as gc
-
+from gbpservice.neutron.nfp.core import periodic_task as core_pt
+from gbpservice.neutron.nfp.config_agent import RestClientOverUnix as rc
+from gbpservice.neutron.nfp.config_agent import loadbalancer as lb
+from gbpservice.neutron.nfp.config_agent import firewall as fw
+from gbpservice.neutron.nfp.config_agent import vpn as vpn
+from gbpservice.neutron.nfp.config_agent import generic as gc
 
 CONFIG_AGENT_MODULES = {'loadbalancer': lb,
                         'firewall': fw,
@@ -25,18 +35,16 @@ class RpcCallback(core_pt.PeriodicTasks):
     def _method_handler(self, rpc_cb):
         if rpc_cb['receiver'] == 'orchestrator':
             mod = CONFIG_AGENT_MODULES['generic']
-            clazz = getattr(mod, 'generic'.title())()
-            method = getattr(clazz, rpc_cb['method'])
-            method(rpc_cb['resource'], **rpc_cb['kwargs'])
+            mod_method = getattr(mod, rpc_cb['method'])
+            mod_method(rpc_cb['resource'], **rpc_cb['kwargs'])
         else:
             mod = CONFIG_AGENT_MODULES[rpc_cb['resource']]
-            clazz = getattr(mod, rpc_cb['resource'].title())()
-            method = getattr(clazz, rpc_cb['method'])
-            method(**rpc_cb['kwargs'])
+            mod_method = getattr(mod, rpc_cb['method'])
+            mod_method(**rpc_cb['kwargs'])
 
     @core_pt.periodic_task(event='PULL_RPC_NOTIFICATIONS', spacing=1)
     def rpc_pull_event(self, ev):
-        rpc_cbs_data = rc.get('nsf/get_notifications')
+        rpc_cbs_data = rc.get('nfp/get_notifications')
         '''
         {response_data : [
             {'receiver': <neutron/orchestrator>,
