+++ gbpservice/neutron/nfp/config_agent/topics.py
@@ -1,0 +1,21 @@
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+
+FW_NFP_CONFIGAGENT_TOPIC = 'oc-firewall-agent'
+LB_NFP_CONFIGAGENT_TOPIC = 'oc-lbaas_agent'
+VPN_NFP_CONFIGAGENT_TOPIC = 'vpn_agent'
+GC_NFP_CONFIGAGENT_TOPIC = 'nfp-dlcm-configurator'
+
+FW_NFP_PLUGIN_TOPIC = 'q-firewall-plugin'
+LB_NFP_PLUGIN_TOPIC = 'n-lbaas-plugin'
+VPN_NFP_PLUGIN_TOPIC = 'vpn_plugin'
+GC_NFP_PLUGIN_TOPIC = 'nfp-configurator-dlcm'
