--- gbpservice/neutron/nsf/config_agent/topics.py
@@ -1,9 +1,0 @@
-FW_NSF_CONFIGAGENT_TOPIC = 'fw_nsf_configagent_topic'
-LB_NSF_CONFIGAGENT_TOPIC = 'lb_nsf_configagent_topic'
-VPN_NSF_CONFIGAGENT_TOPIC = 'vpn_nsf_configagent_topic'
-GC_NSF_CONFIGAGENT_TOPIC = 'gc_nsf_configagent_topic'
-
-FW_NSF_PLUGIN_TOPIC = 'fw_nsf_plugin_topic'
-LB_NSF_PLUGIN_TOPIC = 'lb_nsf_plugin_topic'
-VPN_NSF_PLUGIN_TOPIC = 'vpn_nsf_plugin_topic'
-GC_NSF_PLUGIN_TOPIC = 'gc_nsf_plugin_topic'
