--- gbpservice/neutron/nsf/proxy/cfg.py
@@ -1,5 +1,0 @@
-OPTS = [
-    {'thread_pool_size': 10},
-    {'unix_bind_path': '/tmp/uds_socket'},
-    {'rest_server_address': '0.0.0.0'},
-    {'rest_server_port': '8080'}]
