--- gbpservice/neutron/nsf/proxy/proxy.ini
@@ -1,7 +1,0 @@
-[OPTIONS]
-thread_pool_size= 10
-unix_bind_path= /tmp/uds_socket
-max_connections=10
-rest_server_address= 0.0.0.0
-rest_server_port= 5673
-
